Reorganization (#13)
Shared proto file, gitignore for Qt, style conventions, renaming tree items, starting sprite editor.
