+++ Dialogs/AddImageDialog.ui
@@ -32,9 +32,18 @@
   <property name="modal">
    <bool>true</bool>
   </property>
-  <layout class="QGridLayout" name="gridLayout">
+  <layout class="QFormLayout" name="formLayout">
    <property name="sizeConstraint">
     <enum>QLayout::SetFixedSize</enum>
+   </property>
+   <property name="labelAlignment">
+    <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
+   </property>
+   <property name="horizontalSpacing">
+    <number>4</number>
+   </property>
+   <property name="verticalSpacing">
+    <number>4</number>
    </property>
    <property name="leftMargin">
     <number>4</number>
@@ -48,20 +57,14 @@
    <property name="bottomMargin">
     <number>4</number>
    </property>
-   <property name="spacing">
-    <number>4</number>
-   </property>
-   <item row="5" column="1">
-    <widget class="QDialogButtonBox" name="buttonBox">
-     <property name="orientation">
-      <enum>Qt::Horizontal</enum>
-     </property>
-     <property name="standardButtons">
-      <set>QDialogButtonBox::Cancel|QDialogButtonBox::Ok</set>
+   <item row="0" column="0">
+    <widget class="QLabel" name="widthLabel">
+     <property name="text">
+      <string>Width</string>
      </property>
     </widget>
    </item>
-   <item row="2" column="1">
+   <item row="0" column="1">
     <widget class="QSpinBox" name="widthSpinBox">
      <property name="focusPolicy">
       <enum>Qt::StrongFocus</enum>
@@ -77,21 +80,14 @@
      </property>
     </widget>
    </item>
-   <item row="2" column="0">
-    <widget class="QLabel" name="widthLabel">
-     <property name="text">
-      <string>Width:</string>
-     </property>
-    </widget>
-   </item>
-   <item row="4" column="0">
+   <item row="1" column="0">
     <widget class="QLabel" name="heightLabel">
      <property name="text">
-      <string>Height:</string>
+      <string>Height</string>
      </property>
     </widget>
    </item>
-   <item row="4" column="1">
+   <item row="1" column="1">
     <widget class="QSpinBox" name="heightSpinBox">
      <property name="minimum">
       <number>1</number>
@@ -101,6 +97,16 @@
      </property>
      <property name="value">
       <number>256</number>
+     </property>
+    </widget>
+   </item>
+   <item row="2" column="1">
+    <widget class="QDialogButtonBox" name="buttonBox">
+     <property name="orientation">
+      <enum>Qt::Horizontal</enum>
+     </property>
+     <property name="standardButtons">
+      <set>QDialogButtonBox::Cancel|QDialogButtonBox::Ok</set>
      </property>
     </widget>
    </item>
