+++ Editors/BackgroundEditor.h
@@ -1,45 +1,15 @@
 #ifndef BACKGROUNDEDITOR_H
 #define BACKGROUNDEDITOR_H
 
-#include "ResourceModel.h"
+#include "BaseEdtior.h"
 
-#include "BackgroundRenderer.h"
-
-#include <QDataWidgetMapper>
-#include <QItemDelegate>
-#include <QMetaObject>
-#include <QMetaProperty>
-
-#include <QDebug>
-
-class ImmediateDataWidgetMapper : public QDataWidgetMapper {
-  Q_OBJECT
- private slots:
-  void widgetChanged() {
-	auto delegate = this->itemDelegate();
-	auto editor = static_cast<QWidget *>(this->sender());
-	delegate->commitData(editor);
-	delegate->closeEditor(editor, QAbstractItemDelegate::SubmitModelCache);
-  }
-
- public:
-  ImmediateDataWidgetMapper(QObject *parent = 0) : QDataWidgetMapper(parent) {}
-
-  void addMapping(QWidget *widget, int section, QByteArray propertyName = "") {
-	QDataWidgetMapper::addMapping(widget, section, propertyName);
-	propertyName = this->mappedPropertyName(widget);
-	auto widgetMetaObject = widget->metaObject();
-	auto property = widgetMetaObject->property(widgetMetaObject->indexOfProperty(propertyName));
-	auto metaObject = this->metaObject();
-	QMetaObject::connect(widget, property.notifySignalIndex(), this, metaObject->indexOfSlot("widgetChanged()"));
-  }
-};
+#include "Widgets/BackgroundRenderer.h"
 
 namespace Ui {
 class BackgroundEditor;
 }
 
-class BackgroundEditor : public QWidget {
+class BackgroundEditor : public BaseEditor {
   Q_OBJECT
 
  public:
@@ -48,12 +18,11 @@ class BackgroundEditor : public QWidget {
 
  private slots:
   void dataChanged(const QModelIndex &topLeft, const QModelIndex &bottomRight,
-				   const QVector<int> &roles = QVector<int>());
+                   const QVector<int> &roles = QVector<int>()) override;
   void on_actionSave_triggered();
   void on_actionZoomIn_triggered();
   void on_actionZoomOut_triggered();
   void on_actionZoom_triggered();
-
   void on_actionNewImage_triggered();
 
  private:
