+++ Editors/FontEditor.cpp
@@ -1,7 +1,8 @@
 #include "FontEditor.h"
+
 #include "ui_FontEditor.h"
 
-FontEditor::FontEditor(QWidget* parent, ResourceModel* /*model*/) : QWidget(parent), ui(new Ui::FontEditor) {
+FontEditor::FontEditor(QWidget* parent, ResourceModel* model) : BaseEditor(parent, model), ui(new Ui::FontEditor) {
   ui->setupUi(this);
 }
