+++ Editors/FontEditor.h
@@ -1,15 +1,13 @@
 #ifndef FONTEDITOR_H
 #define FONTEDITOR_H
 
-#include "ResourceModel.h"
-
-#include <QWidget>
+#include "BaseEdtior.h"
 
 namespace Ui {
 class FontEditor;
 }
 
-class FontEditor : public QWidget {
+class FontEditor : public BaseEditor {
   Q_OBJECT
 
  public:
