+++ Editors/PathEditor.cpp
@@ -1,7 +1,8 @@
 #include "PathEditor.h"
+
 #include "ui_PathEditor.h"
 
-PathEditor::PathEditor(QWidget* parent, ResourceModel* /*model*/) : QWidget(parent), ui(new Ui::PathEditor) {
+PathEditor::PathEditor(QWidget* parent, ResourceModel* model) : BaseEditor(parent, model), ui(new Ui::PathEditor) {
   ui->setupUi(this);
 }
