+++ Editors/SpriteEditor.cpp
@@ -1,0 +1,10 @@
+#include "SpriteEditor.h"
+
+#include "ui_SpriteEditor.h"
+
+SpriteEditor::SpriteEditor(QWidget* parent, ResourceModel* model)
+    : BaseEditor(parent, model), ui(new Ui::SpriteEditor) {
+  ui->setupUi(this);
+}
+
+SpriteEditor::~SpriteEditor() { delete ui; }
