+++ Editors/TimelineEditor.cpp
@@ -1,8 +1,9 @@
 #include "TimelineEditor.h"
+
 #include "ui_TimelineEditor.h"
 
-TimelineEditor::TimelineEditor(QWidget* parent, ResourceModel* /*model*/)
-	: QWidget(parent), ui(new Ui::TimelineEditor) {
+TimelineEditor::TimelineEditor(QWidget* parent, ResourceModel* model)
+    : BaseEditor(parent, model), ui(new Ui::TimelineEditor) {
   ui->setupUi(this);
 }
