+++ Models/TreeModel.h
@@ -12,6 +12,7 @@ class TreeModel : public QAbstractItemModel {
   explicit TreeModel(buffers::TreeNode *root, QObject *parent = 0);
   ~TreeModel();
 
+  bool setData(const QModelIndex &index, const QVariant &value, int role) override;
   QVariant data(const QModelIndex &index, int role) const override;
   Qt::ItemFlags flags(const QModelIndex &index) const override;
   QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const override;
