+++ RadialGM.pro.shared
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE QtCreatorProject>
+<qtcreator>
+    <data>
+        <variable>ProjectExplorer.Project.Updater.FileVersion</variable>
+        <value type="int">10</value>
+    </data>
+    <data>
+        <variable>ProjectExplorer.Project.EditorSettings</variable>
+        <valuemap type="QVariantMap">
+            <value type="bool" key="EditorConfiguration.AutoIndent">true</value>
+            <value type="bool" key="EditorConfiguration.AutoSpacesForTabs">true</value>
+            <value type="bool" key="EditorConfiguration.CamelCaseNavigation">true</value>
+            <valuemap type="QVariantMap" key="EditorConfiguration.CodeStyle.0">
+                <value type="QString" key="language">Cpp</value>
+                <valuemap type="QVariantMap" key="value">
+                <value type="QByteArray" key="CurrentPreferences">codestyle</value>
+                <valuelist type="QVariantList" key="UserStickyKeys">
+                    <value type="QString">CurrentPreferences</value>
+                </valuelist>
+                </valuemap>
+            </valuemap>
+            <value type="int" key="EditorConfiguration.CodeStyle.Count">1</value>
+            <value type="QByteArray" key="EditorConfiguration.Codec">UTF-8</value>
+            <value type="bool" key="EditorConfiguration.ConstrainTooltips">true</value>
+            <value type="int" key="EditorConfiguration.IndentSize">2</value>
+            <value type="bool" key="EditorConfiguration.KeyboardTooltips">false</value>
+            <value type="int" key="EditorConfiguration.MarginColumn">120</value>
+            <value type="bool" key="EditorConfiguration.MouseHiding">true</value>
+            <value type="bool" key="EditorConfiguration.MouseNavigation">true</value>
+            <value type="int" key="EditorConfiguration.PaddingMode">1</value>
+            <value type="bool" key="EditorConfiguration.ScrollWheelZooming">true</value>
+            <value type="bool" key="EditorConfiguration.ShowMargin">true</value>
+            <value type="int" key="EditorConfiguration.SmartBackspaceBehavior">0</value>
+            <value type="bool" key="EditorConfiguration.SmartSelectionChanging">true</value>
+            <value type="bool" key="EditorConfiguration.SpacesForTabs">true</value>
+            <value type="int" key="EditorConfiguration.TabKeyBehavior">0</value>
+            <value type="int" key="EditorConfiguration.TabSize">2</value>
+            <value type="bool" key="EditorConfiguration.UseGlobal">false</value>
+            <value type="int" key="EditorConfiguration.Utf8BomBehavior">1</value>
+            <value type="bool" key="EditorConfiguration.addFinalNewLine">true</value>
+            <value type="bool" key="EditorConfiguration.cleanIndentation">true</value>
+            <value type="bool" key="EditorConfiguration.cleanWhitespace">true</value>
+            <value type="bool" key="EditorConfiguration.inEntireDocument">true</value>
+        </valuemap>
+    </data>
+</qtcreator>
